+++ packages/docs/package.json
@@ -23,12 +23,14 @@
 		"@carbon/icons-react": "^11.45.0",
 		"@carbon/react": "^1.61.0",
 		"@stackblitz/sdk": "^1.11.0",
-		"highlight.js": "^11.9.0",
+		"highlight.js": "^11.10.0",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1",
 		"react-router-dom": "^6.24.1"
 	},
 	"devDependencies": {
+		"@netlify/functions": "^2.8.1",
+		"@types/jsdom": "^21.1.7",
 		"@types/react": "^18.3.3",
 		"@types/react-dom": "^18.3.0",
 		"@types/topojson-specification": "^1.0.5",
